# docs/reference/openapi.json
{
  "components": {
    "responses": {
      "Error400": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "The request is invalid; the body says why."
      },
      "Error401": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "No valid credentials (Www-Authenticate: Bearer)."
      },
      "Error403": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "Refused: with no UI auth mode, a client that is not local (only kubectl port-forward is served) or a cross-origin request from an origin not in --cors-allowed-origins; in TokenReview mode, Kubernetes RBAC denied an object the request reads or writes; for the Bundle API, a namespace this controller does not watch."
      },
      "Error404": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "The object does not exist."
      },
      "Error405": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "Method not allowed."
      },
      "Error409": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "Conflict: an ambiguous name, an object that already exists, or nothing to promote or roll back to."
      },
...
      "Error500": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "Internal error."
      },
      "Error503": {
        "content": {
          "text/plain": {
            "schema": {
              "type": "string"
            }
          }
        },
        "description": "TokenReview mode: the TokenReview or SubjectAccessReview API cannot be reached, so the request is refused (fail closed)."
      }
    },
    "schemas": {
...
      "MessageResponse": {
        "additionalProperties": false,
        "description": "uiMessageResponse is the response of an action that creates nothing.",
        "properties": {
          "message": {
            "description": "Message describes what was done.",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "PromoteRequest": {
        "additionalProperties": false,
        "description": "uiPromoteRequest is the body of POST /api/v1/ui/promote.",
        "properties": {
          "environment": {
            "description": "Environment is the environment to promote to; its upstreams must have a Verified Bundle.",
            "type": "string"
          },
          "namespace": {
            "description": "Namespace of the Pipeline. Defaults to \"default\".",
            "type": "string"
          },
          "pipeline": {
            "description": "Pipeline is the Pipeline to promote in.",
            "type": "string"
          }
        },
        "required": [
          "environment",
          "pipeline"
        ],
        "type": "object"
      },
      "PromoteResponse": {
        "additionalProperties": false,
        "description": "uiPromoteResponse is the 201 response of POST /api/v1/ui/promote.",
        "properties": {
          "bundle": {
            "description": "Bundle is the promotion Bundle created.",
            "type": "string"
          },
          "message": {
            "description": "Message describes the promotion.",
            "type": "string"
          },
          "source": {
            "description": "Source is the Verified Bundle whose artifacts it promotes.",
            "type": "string"
          }
        },
        "required": [
          "bundle",
          "message",
          "source"
        ],
        "type": "object"
      },
...
      "RollbackRequest": {
        "additionalProperties": false,
        "description": "uiRollbackRequest is the body of POST /api/v1/ui/rollback.",
        "properties": {
          "environment": {
            "description": "Environment is the environment to roll back.",
            "type": "string"
          },
          "hold": {
            "description": "Hold keeps the environment on the rollback until it is released (kardinal rollback --hold): no other Bundle promotes into it, and the rollback passes the gates that would block it, each pass audited.",
            "type": "boolean"
          },
          "holdReason": {
            "description": "HoldReason says why the environment is held. Required with hold.",
            "type": "string"
          },
          "namespace": {
            "description": "Namespace of the Pipeline. Defaults to \"default\".",
            "type": "string"
          },
          "pipeline": {
            "description": "Pipeline is the Pipeline to roll back.",
            "type": "string"
          },
          "toBundle": {
            "description": "ToBundle is the Bundle to go back to. Empty means the most recent Bundle, other than the one deployed now, that was Verified in the environment.",
            "type": "string"
          }
        },
        "required": [
          "environment",
          "pipeline"
        ],
        "type": "object"
      },
      "RollbackResponse": {
        "additionalProperties": false,
        "description": "uiRollbackResponse is the 201 response of POST /api/v1/ui/rollback.",
        "properties": {
          "bundle": {
            "description": "Bundle is the rollback Bundle created.",
            "type": "string"
          },
          "held": {
            "description": "Held is true when the environment is held on the rollback.",
            "type": "boolean"
          },
          "message": {
            "description": "Message describes the rollback.",
            "type": "string"
          },
          "rollbackOf": {
            "description": "RollbackOf is the Bundle whose artifacts it restores.",
            "type": "string"
          }
        },
        "required": [
          "bundle",
          "message",
          "rollbackOf"
        ],
        "type": "object"
      },
...
      "uiReleaseHoldRequest": {
        "additionalProperties": false,
        "description": "uiReleaseHoldRequest is the body of POST /api/v1/ui/release-hold.",
        "properties": {
          "environment": {
            "description": "Environment is the held environment.",
            "type": "string"
          },
          "namespace": {
            "description": "Namespace of the Pipeline. Defaults to \"default\".",
            "type": "string"
          },
          "pipeline": {
            "description": "Pipeline is the Pipeline whose environment is held.",
            "type": "string"
          }
        },
        "required": [
          "environment",
          "pipeline"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bundleToken": {
        "description": "The Bundle API token (bundleAPI.tokenSecretRef, --bundle-api-token).",
        "scheme": "bearer",
        "type": "http"
      },
      "uiToken": {
        "description": "The static UI token (ui.auth.tokenSecretRef) or, with ui.auth.tokenReview, a Kubernetes token (kubectl create token) whose user Kubernetes RBAC authorizes for each request. With neither mode set, the UI API takes no credentials and answers only loopback peers (kubectl port-forward).",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The UI API (/api/v1/ui/*, on the UI listener), the Bundle API for CI (POST /api/v1/bundles) and the SCM webhook health endpoint (on the webhook listener). Authentication and RBAC: https://pnz1990.github.io/kardinal-promoter/reference/rest-api/",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "kardinal-promoter REST API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ui/promote": {
      "post": {
        "operationId": "promote",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromoteResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "security": [
          {
            "uiToken": []
          },
          {}
        ],
        "summary": "Promote the Bundle Verified upstream into an environment (kardinal promote)",
        "tags": [
          "UI"
        ]
      }
    },
    "/api/v1/ui/release-hold": {
      "post": {
        "operationId": "releaseHold",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/uiReleaseHoldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "security": [
          {
            "uiToken": []
          },
          {}
        ],
        "summary": "Release the hold of a rollback on an environment (kardinal release-hold)",
        "tags": [
          "UI"
        ]
      }
    },
...
    "/api/v1/ui/rollback": {
      "post": {
        "operationId": "rollback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "405": {
            "$ref": "#/components/responses/Error405"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          }
        },
        "security": [
          {
            "uiToken": []
          },
          {}
        ],
        "summary": "Roll an environment back to an earlier Verified Bundle (kardinal rollback)",
        "tags": [
          "UI"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "The UI listener (--ui-listen-address). https:// when controller TLS is set.",
      "url": "http://{controller}:8082",
      "variables": {
        "controller": {
          "default": "kardinal-promoter.kardinal-system.svc"
        }
      }
    }
  ]
}